refactor(proxy): split model_gone.go and upstream_classify.go by concern
Same package, pure moves with their doc comments:

- model_gone.go (1163 -> 690) keeps the thresholds, noteModelGone,
  probeForRetirement and acquireProbeSlot.
  model_gone_streak.go takes goneStreakKey, goneStreak and its methods
  (strike, claimProbe, canClaimProbe, park, supersede, clearLocked,
  noteInconclusiveProbe, count); model_gone_modality.go takes the
  modality predicates; model_gone_stream.go takes streamVerdict with
  its constants, verdictForStream, producedOutput, noteStreamOutcome
  and noteModelServed.
- upstream_classify.go (804 -> 410) keeps the structured-error parsing
  and classifyUpstreamError; upstream_classify_phrase.go takes the
  prose analysis (the verb table, the model-ID scanner with its version
  suffix and vendor prefix rules, the clause-gap attribution and
  phraseIsAbout).

Both entries leave the size allowlist. Tests are unchanged.
