+++ internal/proxy/model_gone.go
@@ -2,11 +2,8 @@ package proxy
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"slices"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"github.com/google/uuid"
@@ -15,7 +12,6 @@ import (
 	"github.com/hugalafutro/model-hotel/internal/events"
 	"github.com/hugalafutro/model-hotel/internal/failover"
 	"github.com/hugalafutro/model-hotel/internal/metrics"
-	"github.com/hugalafutro/model-hotel/internal/model"
 )
 
 // goneStrikeThreshold is how many consecutive KindProviderModelGone responses a
@@ -127,317 +123,6 @@ const goneProbeMaxConcurrent = 4
 // exactly when an undersized group most needs resizing. Nothing on the request
 // path is waiting on the total.
 const goneWriteTimeout = 10 * time.Second
-
-// goneStreakKey identifies a streak: one model, on one upstream surface.
-//
-// The surface is the PROBE endpoint (see probeEndpointForFamily), not the
-// endpoint family, so chat and messages share a streak while embeddings keeps
-// its own. That is the right grain because it is exactly what the probe can ask:
-// two front doors onto the same question are one question, and a different door
-// onto a different question is a different streak.
-type goneStreakKey struct {
-	model    uuid.UUID
-	endpoint string
-}
-
-// modalityRulesOutSurface reports whether a gone-classified refusal that arrived
-// on this upstream surface may count against the model at all.
-//
-// Nothing filters a request by modality on the way in, so `POST /v1/embeddings`
-// naming a chat model is forwarded, and a provider answering "gpt-4o is not
-// supported for embeddings" has named the model beside a gone-phrase. The probe
-// cannot rescue it: it asks on the surface the strikes arrived on, so it
-// reproduces the misuse and confirms.
-//
-// The two surfaces answer to different burdens of proof, because the cost of
-// guessing is not symmetric:
-//
-//   - Embeddings requires POSITIVE evidence — the catalog has to say this model
-//     produces embeddings. Being wrong retires a live chat model gateway-wide;
-//     being cautious only means an embeddings model whose catalog declares
-//     nothing is never auto-retired, the same trade probeEndpointForFamily takes
-//     for rerank. It matters because liveModelStub writes "[]" for every model no
-//     catalog covers.
-//   - Chat keeps the opposite default. Chat is what most models are and what
-//     most refusals arrive on, so demanding a declared modality would switch
-//     traffic-driven retirement off for every uncatalogued model at once. A
-//     silent catalog may strike; one that positively describes something a chat
-//     completion cannot be about may not.
-//
-// The chat test reads both arrays because either can settle it and neither can
-// alone: an image model declares output ["image"] and a TTS model ["audio"],
-// while a speech-to-text model declares output ["text"] like any chat model and
-// gives itself away on the INPUT side with ["audio"].
-//
-// Text and code are the whole allow-list, and deliberately not the vocabulary of
-// things a chat completion can CARRY: a chat model that also emits images
-// declares ["text","image"] and is admitted by the text, while one declaring
-// image alone is an images-endpoint model whatever it can be coaxed into. See
-// canonicalModalityRank in internal/provider for the full vocabulary.
-func modalityRulesOutSurface(m *model.Model, probeEndpoint string) bool {
-	out := declaredModalities(m.OutputModalities)
-	if probeEndpoint == probeEmbeddingsEndpoint {
-		// Positive evidence, so an undeclared list rules the surface out.
-		return !slices.Contains(out, "embedding")
-	}
-	return !admitsChatModalities(out) || !admitsChatModalities(declaredModalities(m.InputModalities))
-}
-
-// modalityAdmitsBothProbeSurfaces reports whether the catalog says this model
-// serves chat AND embeddings.
-//
-// Such a model is not auto-retired at all, because of the mismatch between what
-// the evidence is about and what the action does. Strikes, probes and successes
-// are per surface, but AutoRetireIfConfirmed disables the model ROW, so a chat
-// retirement would take the working embeddings surface with it. No probe can
-// catch that: the probe is right, the model really is gone on chat, and the
-// disable is simply broader than the finding.
-//
-// Retiring per surface is the honest fix and the schema does not offer it — one
-// row is one model, enabled or not — so this takes the same trade as image, TTS,
-// STT and rerank and leaves the model enabled until discovery drops it or an
-// operator disables it by hand.
-//
-// It costs almost nothing: it needs a provider serving one model id on both
-// surfaces, which is rare, and an empty modality list already rules the
-// embeddings surface out.
-func modalityAdmitsBothProbeSurfaces(m *model.Model) bool {
-	return !modalityRulesOutSurface(m, probeChatEndpoint) && !modalityRulesOutSurface(m, probeEmbeddingsEndpoint)
-}
-
-// declaredModalities parses one of the model's modality columns, returning nil
-// for anything it cannot read as a list. Callers treat nil as "the catalog says
-// nothing", which is not the same as "the catalog says no".
-func declaredModalities(raw string) []string {
-	if raw == "" || raw == "[]" {
-		return nil
-	}
-	var list []string
-	if json.Unmarshal([]byte(raw), &list) != nil {
-		return nil
-	}
-	return list
-}
-
-// admitsChatModalities reports whether a declared modality list leaves room for
-// a chat completion. An empty list does: nothing declared is not evidence.
-func admitsChatModalities(list []string) bool {
-	if len(list) == 0 {
-		return true
-	}
-	return slices.Contains(list, "text") || slices.Contains(list, "code")
-}
-
-// goneStreak is one model's consecutive-refusal count plus a tombstone for the
-// window between deciding to disable and the write landing.
-//
-// The counters are guarded by mu rather than being separate atomics, because
-// none of the three decisions can be expressed atomically: deciding whether the
-// strike window has lapsed and then applying the decision is one operation, and
-// splitting it loses strikes (a reset racing two increments stores 1 after both
-// have added). nextProbeAt is the same shape — read the deadline, decide, stamp
-// the new one — and splitting it would let two callers past the same expired
-// deadline and issue two probes where the point is to issue one. The lock is
-// held for a comparison and an addition on a per-model struct that is contended
-// only while that model is failing.
-//
-// cancelled is a tombstone for the window between deciding to disable and the
-// write landing, during which the model can answer a request and prove it is
-// alive. It stays atomic because the detached disable goroutine reads it while
-// the request path may be writing it. It is read twice and both reads are
-// needed: before the write to prevent a disable now known to be wrong, and after
-// it to catch a success that landed while the write was in flight — which the
-// first read is by definition too early to see, and which can only be undone
-// rather than prevented.
-type goneStreak struct {
-	mu           sync.Mutex
-	n            int64
-	lastStrike   time.Time
-	nextProbeAt  time.Time
-	inconclusive int
-
-	cancelled atomic.Bool
-}
-
-// strike records a refusal and returns the length of the streak it belongs to.
-//
-// A strike arriving more than goneStrikeWindow after the previous one starts the
-// count over instead of extending it, so a retirement is always drawn from one
-// run of recent traffic rather than from unrelated failures that happen to share
-// a model.
-func (s *goneStreak) strike(now time.Time) int64 {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if !s.lastStrike.IsZero() && now.Sub(s.lastStrike) > goneStrikeWindow {
-		s.n = 1
-	} else {
-		s.n++
-	}
-	s.lastStrike = now
-	return s.n
-}
-
-// claimProbe reports whether this caller may spend an upstream request on the
-// model, and takes the right to do so if it may.
-//
-// It is one operation and not two, which is what makes it serve both jobs.
-// Claiming at spawn time rather than crediting at completion means a second
-// caller arriving while the first probe is still in flight finds the stamp
-// already set and drops out, so this is also the "a probe is already in flight"
-// guard — hence goneProbeCooldown's floor being the probe's own deadline. A
-// burst of concurrent refusals therefore issues a single probe, and a parked
-// streak stays re-probeable by the next refusal past the cooldown, so postponing
-// does not have to throw the evidence away to stay reachable.
-//
-// A zero stamp admits the first caller, which is what a model that has never
-// been probed should get.
-//
-// The count check is the other half of the same critical section and not a
-// formality. The caller strikes and then claims, and those are two lock
-// acquisitions: a success can land in between, clear the count and set the
-// tombstone, and the claim would otherwise spend an upstream request on strikes
-// that no longer exist.
-//
-// Granting a claim also clears the tombstone, which is what makes a streak
-// reusable. The flag means "a success landed after the disable now in flight was
-// decided", so it belongs to one decision; a claim starts the next one, and the
-// count check guarantees the strikes it starts on are newer than any success
-// that set the flag. Leaving it set would make the disable this claim spawns
-// stand down at its own pre-write check, and the model would never be retired
-// again on this streak. Nothing can be racing an earlier probe: goneProbeCooldown
-// is five minutes and the whole goroutine lives at most goneProbeTimeout plus
-// one write.
-func (s *goneStreak) claimProbe(now time.Time) bool {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.n < goneStrikeThreshold {
-		return false
-	}
-	if !s.nextProbeAt.IsZero() && now.Before(s.nextProbeAt) {
-		return false
-	}
-	s.nextProbeAt = now.Add(goneProbeCooldown)
-	s.cancelled.Store(false)
-	return true
-}
-
-// canClaimProbe answers the same question as claimProbe without taking anything.
-//
-// It is not a substitute for the claim and cannot be: two callers can both read
-// true and only one may proceed. It exists so the cheap, per-model reason to
-// stop is checked before the shared, per-provider one, which is what keeps the
-// semaphore counting adjudications rather than refusals — without it, a retry
-// storm across models that are all inside their cooldowns holds all four slots
-// for the microseconds each take-and-release costs, denying a genuine probe to
-// the one model whose cooldown HAS expired.
-//
-// It mirrors both of claimProbe's conditions, including the count: reading only
-// the stamp would let a refusal whose evidence a success had already cleared go
-// take a provider slot, to be turned away by the claim a moment later.
-//
-// The reason is returned because the caller logs it and the two cases are not
-// the same event: a cooldown is a deliberate wait, while a count below the
-// threshold means a success cleared the evidence between the strike and this
-// read.
-func (s *goneStreak) canClaimProbe(now time.Time) (ok bool, reason string) {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.n < goneStrikeThreshold {
-		return false, "the strikes were cleared while this refusal was being recorded"
-	}
-	if !s.nextProbeAt.IsZero() && now.Before(s.nextProbeAt) {
-		return false, "the model is inside its probe cooldown"
-	}
-	return true, ""
-}
-
-// park clears the evidence a streak has accumulated and keeps the probe
-// cooldown it has earned. It is what "clear the streak" means everywhere in
-// this file; nothing removes the entry.
-//
-// Deleting the entry would silently discard the rate bound: nextProbeAt lives on
-// the streak, so the next refusal would build a fresh zero-valued one whose
-// claimProbe admits immediately. Every reason a streak is cleared is a reason it
-// may be rebuilt at once — a probe that answered while traffic keeps drawing
-// retirement prose, a success between refusals, a disable that failed to write —
-// so each would turn into three fresh refusals buying another upstream call,
-// forever.
-//
-// Parking keeps both halves of the bound: the count is reset, so the model needs
-// three FRESH refusals, and the stamp survives, so the reconsideration also
-// waits out the cooldown.
-//
-// Mutating in place rather than reaching into h.goneStrikes by model id makes
-// every caller identity-scoped for free: sync.Map holds the pointer, so a park
-// lands on the map entry while this is still the live streak.
-//
-// The tombstone is deliberately not touched: a success sets it (noteModelServed)
-// and a claim clears it (claimProbe), and both are decisions about a disable
-// rather than about the evidence.
-//
-// Nothing is ever deleted, so the map holds one small struct per model that has
-// ever drawn a gone-classified refusal — bounded by the catalog, and the point:
-// a model's probe cooldown has to outlive the streak that earned it.
-func (s *goneStreak) park() {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.clearLocked()
-}
-
-// supersede is park plus the tombstone: the model answered, so a disable that
-// has been decided stands down and the evidence behind it is cleared.
-//
-// One critical section, which is why this is not two calls at the call site. The
-// disable goroutine reads the tombstone without the lock and then asks for the
-// count, so the two updates have to be indivisible from its point of view:
-// seeing cancelled set while the count still shows the strikes that caused this
-// retirement reads as "the model is refusing again", the revert is skipped, and
-// the count is parked at zero immediately afterwards so nothing triggers one
-// again — leaving a model that has just answered disabled until an operator
-// re-enables it by hand. Storing inside the lock closes that: once a reader
-// observes the tombstone the writer still holds mu.
-//
-// It reports whether it changed anything, for the caller's log line rather than
-// for control flow. The early return is deliberately narrow — a count of zero
-// with no tombstone can still belong to a disable this success has to cancel.
-func (s *goneStreak) supersede() bool {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.n == 0 && s.cancelled.Load() {
-		return false
-	}
-	s.cancelled.Store(true)
-	s.clearLocked()
-	return true
-}
-
-// clearLocked resets the evidence. Callers hold mu.
-//
-// The inconclusive run goes with it: it counts probes that could not answer a
-// question, so clearing the question ends the run. Every caller is the model
-// answering — a success, or a probe that got content out of it — which is the
-// outcome the run was waiting for.
-func (s *goneStreak) clearLocked() {
-	s.n = 0
-	s.lastStrike = time.Time{}
-	s.inconclusive = 0
-}
-
-// noteInconclusiveProbe records a probe that established nothing and reports how
-// many in a row this streak has now taken. See goneProbeInconclusiveWarnAfter.
-func (s *goneStreak) noteInconclusiveProbe() int {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	s.inconclusive++
-	return s.inconclusive
-}
-
-// count reports the current streak length.
-func (s *goneStreak) count() int64 {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.n
-}
 
 // noteModelGone records one strike against a model the provider refused as
 // retired, and once the streak reaches goneStrikeThreshold asks the provider
@@ -1001,163 +686,5 @@ func (h *Handler) acquireProbeSlot(providerID uuid.UUID) (release func(), ok boo
 		return sync.OnceFunc(func() { <-sem }), true
 	default:
 		return nil, false
-	}
-}
-
-// streamVerdict is what a finished stream says about whether the model still
-// exists. Three-way on purpose, and the middle case is the one that matters.
-type streamVerdict int
-
-const (
-	// verdictInconclusive: the stream failed for a reason unrelated to the
-	// model's existence. Leave the strike streak untouched.
-	verdictInconclusive streamVerdict = iota
-	// verdictGone: the provider reported the model retired mid-stream.
-	verdictGone
-	// verdictServed: the stream completed cleanly, so the model answered.
-	verdictServed
-)
-
-// verdictForStream maps a finished stream to what it proves about the model.
-//
-// The trap this encodes: treating "not gone" as "served" lets a retired model
-// stay routable forever, because its own unrelated failures (transient provider
-// errors, client disconnects, stalls) keep resetting the count that would have
-// retired it. Equally, treating any failure as evidence of death would disable
-// a healthy model during an outage.
-//
-// producedOutput is the second half of that care. An absent error kind is not by
-// itself proof the model answered: a stream can open, emit nothing and end
-// without ever recording an error, and crediting that as a success would clear a
-// retirement streak on the strength of nothing at all. Clearing therefore
-// requires positive evidence that content actually flowed.
-// upstreamKind is what the provider said, which is the only thing that can
-// establish a retirement. kind is how the request ended, and it is what rules a
-// success out.
-//
-// The two are separate because they answer different questions and the second
-// overwrites the first. A provider can report the model retired mid-stream and
-// the client can then hang up — the likeliest thing for a client to do on
-// receiving an error — at which point the recorded kind becomes
-// client_disconnect. Judging the model by that would let a client suppress the
-// evidence by reacting to it, and a retired model would stay routable for as
-// long as clients kept disconnecting on its errors.
-func verdictForStream(kind, upstreamKind ErrorKind, producedOutput bool) streamVerdict {
-	switch {
-	case upstreamKind == KindProviderModelGone || kind == KindProviderModelGone:
-		return verdictGone
-	case kind == "" && producedOutput:
-		return verdictServed
-	default:
-		return verdictInconclusive
-	}
-}
-
-// producedOutput reports whether a response actually delivered content.
-//
-// Used by the stream verdict and by the non-streaming chat path, which ask the
-// same question of the same fields: a 200 is a status, and what clears a
-// retirement streak has to be an answer.
-//
-// deliveredContent is the authoritative signal, set where the content itself is
-// observed. The other two are corroboration and neither can be relied on alone:
-// completion tokens are absent when a provider omits the usage chunk, and TTFT
-// is zero when the probe is switched off. With only those two, a provider that
-// streams a perfectly good answer and reports no usage, on a gateway with the
-// probe off, reads as having produced nothing — the success then fails to clear
-// the streak, and later refusals retire a model whose failures were never
-// consecutive.
-//
-// The direction of the error matters here, which is why the bar is positive
-// evidence rather than absence of an error: a stream can open, emit nothing and
-// end without recording anything, and crediting that would clear a retirement
-// streak on the strength of an empty response.
-func producedOutput(logData *requestLogData) bool {
-	if logData == nil {
-		return false
-	}
-	return logData.deliveredContent || logData.tokensCompletion > 0 || logData.ttftMs > 0
-}
-
-// noteStreamOutcome applies the model verdict once a stream has finished. Shared
-// by the sequential dispatch and the hedged winner so the two cannot drift —
-// the hedged path previously returned without recording any verdict at all, so
-// a model retired mid-stream stayed routable whenever hedging was enabled.
-func (h *Handler) noteStreamOutcome(logData *requestLogData, candidate modelCandidate) {
-	// The guard belongs here, not in producedOutput. That function checks for nil
-	// too, but this expression dereferences logData to build its arguments and Go
-	// evaluates all of them before the call, so the helper's guard could never
-	// run. Same shape as probeForRetirement's.
-	if logData == nil {
-		return
-	}
-	switch verdictForStream(logData.errorKind, logData.upstreamKind, producedOutput(logData)) {
-	case verdictGone:
-		// The endpoint family comes off the log entry rather than being assumed:
-		// it is what decides whether this refusal can be adjudicated at all, and
-		// a stream reaching here can have started at /v1/chat/completions or at
-		// /v1/messages. newPendingRequestLog stamps it at ingest, so it is set on
-		// every path that can reach this.
-		h.noteModelGone(candidate, logData.endpointType)
-	case verdictServed:
-		h.noteModelServed(candidate.model, logData.endpointType)
-	case verdictInconclusive:
-		// Deliberately nothing: see verdictForStream.
-	}
-}
-
-// noteModelServed clears any accumulated gone-strikes after the model answers.
-// The strike streak must be consecutive, so one success is enough to reset it.
-//
-// It runs on the request path, so what it costs there is part of the design: a
-// model that has never been refused misses the map entirely, and one that has
-// pays a miss plus an uncontended mutex on a request that is already waiting on
-// an upstream call. Nothing is written and nothing touches the database.
-//
-// It also cancels a disable that has been decided but not yet written, which is
-// why it goes through supersede rather than setting the flag and parking
-// separately: a queued disable reads both and has to see them change together.
-//
-// The entry is parked, not dropped, and the difference is the probe cooldown.
-// Deleting it took nextProbeAt with it, so a success between refusals reset the
-// rate bound — and a model that refuses some request shapes while serving others
-// produces exactly that interleaving. A success clears what the model is accused
-// of; it does not buy the gateway another free upstream call. See park.
-//
-// The success clears ONE surface: the one it arrived on. Clearing globally
-// answered two separate questions as one — a provider that had retired a model's
-// chat surface while still serving its embeddings would have every embeddings
-// success wipe the chat streak, so the dead surface could never accumulate three
-// consecutive strikes and would never be adjudicated at all. Tightening it is
-// safe because what retires a model is the PROBE, and the probe asks on the same
-// surface.
-//
-// A family that cannot be probed clears nothing, following the strike side: an
-// image or TTS response is not evidence about the chat surface any more than an
-// image refusal is. It is also why this takes an endpointType rather than
-// deriving one — every caller has the family that ingest stamped.
-//
-// The log line is conditional on supersede having changed something. Nothing
-// removes entries, so a model that drew one refusal at 09:00 keeps its parked
-// streak for the life of the process, and an unconditional line would claim to
-// have "cleared gone-strikes" on every successful request from then on.
-func (h *Handler) noteModelServed(m *model.Model, endpointType string) {
-	if m == nil || m.ID == uuid.Nil {
-		return
-	}
-	endpoint, ok := probeEndpointForFamily(endpointType)
-	if !ok {
-		return
-	}
-	raw, ok := h.goneStrikes.Load(goneStreakKey{model: m.ID, endpoint: endpoint})
-	if !ok {
-		return
-	}
-	streak, ok := raw.(*goneStreak)
-	if !ok {
-		return
-	}
-	if streak.supersede() {
-		debuglog.Debug("proxy: model answered again, cleared gone-strikes", "model", m.ModelID, "endpoint", endpoint)
 	}
 }
